10787: Combine then extract FullAddressSearch component (#9179)
* changelog: Internal, In-Person Proofing, Separate components

* Move Into separate component

* Lint
