+++ app/javascript/packages/document-capture/components/in-person-location-full-address-entry-post-office-search-step.tsx
@@ -1,10 +1,8 @@
 import { useState, useEffect, useCallback, useRef, useContext } from 'react';
-import { t } from '@18f/identity-i18n';
-import { Alert, PageHeading } from '@18f/identity-components';
 import { request } from '@18f/identity-request';
 import { forceRedirect } from '@18f/identity-url';
-import { transformKeys, snakeCase, InPersonLocations } from '@18f/identity-address-search';
-import type { LocationQuery, FormattedLocation } from '@18f/identity-address-search/types';
+import { transformKeys, snakeCase } from '@18f/identity-address-search';
+import type { FormattedLocation } from '@18f/identity-address-search/types';
 import FullAddressSearch from './in-person-full-address-search';
 import BackButton from './back-button';
 import AnalyticsContext from '../context/analytics';
@@ -19,24 +17,13 @@ function InPersonLocationFullAddressEntryPostOfficeSearchStep({
 }) {
   const { inPersonURL } = useContext(InPersonContext);
   const [inProgress, setInProgress] = useState<boolean>(false);
-  const [isLoadingLocations, setLoadingLocations] = useState<boolean>(false);
   const [autoSubmit, setAutoSubmit] = useState<boolean>(false);
   const { trackEvent } = useContext(AnalyticsContext);
   const [locationResults, setLocationResults] = useState<FormattedLocation[] | null | undefined>(
     null,
   );
-  const [foundAddress, setFoundAddress] = useState<LocationQuery | null>(null);
-  const [apiError, setApiError] = useState<Error | null>(null);
   const [disabledAddressSearch, setDisabledAddressSearch] = useState<boolean>(false);
   const { flowPath } = useContext(UploadContext);
-
-  const onFoundLocations = (
-    address: LocationQuery | null,
-    locations: FormattedLocation[] | null | undefined,
-  ) => {
-    setFoundAddress(address);
-    setLocationResults(locations);
-  };
 
   // ref allows us to avoid a memory leak
   const mountedRef = useRef(false);
@@ -105,28 +92,13 @@ function InPersonLocationFullAddressEntryPostOfficeSearchStep({
 
   return (
     <>
-      {apiError && (
-        <Alert type="error" className="margin-bottom-4">
-          {t('idv.failure.exceptions.post_office_search_error')}
-        </Alert>
-      )}
-      <PageHeading>{t('in_person_proofing.headings.po_search.location')}</PageHeading>
-      <p>{t('in_person_proofing.body.location.po_search.po_search_about')}</p>
       <FullAddressSearch
         registerField={registerField}
-        onFoundLocations={onFoundLocations}
-        onLoadingLocations={setLoadingLocations}
-        onError={setApiError}
+        onFoundLocations={setLocationResults}
         disabled={disabledAddressSearch}
         locationsURL={LOCATIONS_URL}
+        handleLocationSelect={handleLocationSelect}
       />
-      {locationResults && foundAddress && !isLoadingLocations && (
-        <InPersonLocations
-          locations={locationResults}
-          onSelect={handleLocationSelect}
-          address={foundAddress.address || ''}
-        />
-      )}
       <BackButton role="link" includeBorder onClick={toPreviousStep} />
     </>
   );
